refactor: extract ComposeUpdatesController from ComposeController (PR6)
First seam of the ComposeController (1902-line god class) split. Moves the six
image-update endpoints into a focused ComposeUpdatesController:

  GET  projects/{name}/check-updates
  POST projects/{name}/update
  GET  update-status
  POST update-all
  POST clear-update-cache
  POST check-all-updates

The new controller declares an explicit [Route("api/compose")] so the public
routes are byte-for-byte unchanged (frontend untouched). It depends only on the
four services those endpoints actually use (IComposeUpdateService, IAuditService,
ISelfFilterService, ILogger) instead of ComposeController's ~18 dependencies;
IComposeUpdateService is no longer injected into ComposeController.

ComposeController: 1902 -> 1631 lines. Behavior preserved (endpoints moved verbatim).

Verify: dotnet build -c Release -> 0 warnings/errors; dotnet test -> 305 passed;
app boots with no ambiguous-route error and every moved/remaining endpoint routes
(401 auth challenge, not 404/500).
